projects/H-032/requirements.txt
flask
torch
transformers
requests
pillow
matplotlib